Separando callbacks do caption_page
